feat: move add panel code from inner component to parent component
